d23apiv1-nikhil-dev/requirements.txt
# d23.ai WhatsApp AI Assistant Dependencies
# ==========================================

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6



# LangChain / LangGraph
langchain==1.2.0
langchain-core==1.2.1
langchain-openai==1.1.3
langgraph==1.0.5
langchain-community==0.4.1
langchain-text-splitters==1.1.0
langchain-chroma==1.1.0

# HTTP Client
httpx>=0.25.0

# AI/ML APIs
tavily-python>=0.3.0
fal-client>=0.4.0

# Text-to-Speech (Edge TTS - free, no API key needed)
edge-tts>=7.0.0

# Speech-to-Text (MLX Whisper - local, free, optimized for Apple Silicon)
mlx-whisper>=0.4.0

# Configuration & Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0


langsmith>=0.0.87

# Async Support
anyio>=3.0.0

# Development (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0

# Chainlit
chainlit>=1.0.0

# Utilities
fuzzywuzzy>=0.18.0
pypdf>=4.0.0
chromadb>=0.4.0

tiktoken>=0.7.0,<1.0.0
psycopg2-binary>=2.9.0
apscheduler>=3.10.0
parsedatetime>=2.6
pytz>=2024.1
psycopg-pool>=3.2.0
langgraph-checkpoint-postgres>=2.0.0


